# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Retail app Api",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "deendipo@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://retail-app.test/api",
            "description": " API Documentation Server"
        }
    ],
    "paths": {
...
        "/inventory": {
            "get": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Get list of Inventories",
                "description": "Returns list of Inventories",
                "operationId": "getInventoryList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "bolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success."
                                        },
                                        "data": {
                                            "description": "bla bla bla",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Package One"
                                                    },
                                                    "price": {
                                                        "type": "number",
                                                        "example": "20000"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2022-07-10T19:16:29.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2022-07-10T19:16:29.000000Z"
                                                    },
                                                    "inventory_quantity": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "number",
                                                                "example": "18"
                                                            },
                                                            "inventory_id": {
                                                                "type": "number",
                                                                "example": "1"
                                                            },
                                                            "quantity": {
                                                                "type": "number",
                                                                "example": "200"
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "example": "2022-07-10T19:16:29.000000Z"
                                                            },
                                                            "updated_at": {
                                                                "type": "string",
                                                                "example": "2022-07-10T19:16:29.000000Z"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Store new inventory",
                "description": "Returns inventory data",
                "operationId": "storeInventory",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Inventory"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inventory created successful.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "bolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Inventory created successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "number",
                                                    "example": "1"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Package One"
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "example": "20000"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2022-07-10T19:16:29.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2022-07-10T19:16:29.000000Z"
                                                },
                                                "inventory_quantity": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "number",
                                                            "example": "18"
                                                        },
                                                        "inventory_id": {
                                                            "type": "number",
                                                            "example": "1"
                                                        },
                                                        "quantity": {
                                                            "type": "number",
                                                            "example": "200"
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "example": "2022-07-10T19:16:29.000000Z"
                                                        },
                                                        "updated_at": {
                                                            "type": "string",
                                                            "example": "2022-07-10T19:16:29.000000Z"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                }
            }
        },
...
        "/inventory/delete/{id}": {
            "delete": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Delete existing Inventory",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteInventory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Inventory id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "Bolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Inventory destroyed successfully."
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "405": {
                        "description": "Unable to delete Inventory"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Inventory": {
                "title": "Inventory",
                "description": "Inventory model",
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name of the new inventory",
                        "type": "string",
                        "example": "Package One"
                    },
                    "price": {
                        "title": "price",
                        "description": "Inventory price",
                        "type": "number",
                        "example": "20000"
                    },
                    "quantity": {
                        "title": "quantity",
                        "description": "Inventory quantity",
                        "type": "string",
                        "example": "200"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Inventory"
                }
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}